+++ console/jest.config.js
@@ -2,8 +2,10 @@ module.exports = {
   roots: ['<rootDir>/src'],
   transform: {
     '^.+\\.(ts|tsx|js|jsx)?$': 'ts-jest',
+    '^.+\\.svg$': 'jest-svg-transformer',
   },
   testRegex: '(/__tests__/.*)\\.(test|spec).[jt]sx?$',
+  setupFilesAfterEnv: ['<rootDir>/src/setupTests.ts'],
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'json', 'node'],
   moduleDirectories: ['node_modules', 'src'],
   testEnvironment: 'jsdom',