+++ console/src/types.ts
@@ -73,6 +73,9 @@ export type RunSqlType = {
   };
 };
 
+export type Entry<O, K extends keyof O> = [K, O[K]];
+export type Entries<O> = Entry<O, keyof O>[];
+
 declare global {
   // eslint-disable-next-line no-underscore-dangle
   const __DEVELOPMENT__: boolean;