+++ console/src/utils/migration/utils.ts
@@ -356,15 +356,16 @@ export const getDownQueryComments = (
 ) => {
   if (Array.isArray(upqueries) && upqueries.length >= 0) {
     const comment = [
-      "Could not auto-generate a down migration.",
-      "Please write an appropriate down migration for the SQL below:",
-      ...upqueries.map((i) => i.args.sql),
-      ""
-    ].join("\n")
-     // Normalize \r\n to \n and add comments before every line
-     .replace(/\r?(^|\n)(?!$)/g, "$1-- ")
-     // Eliminate trailing spaces
-     .replace(/ +\n/g, "\n");
+      'Could not auto-generate a down migration.',
+      'Please write an appropriate down migration for the SQL below:',
+      ...upqueries.map(i => i.args.sql),
+      '',
+    ]
+      .join('\n')
+      // Normalize \r\n to \n and add comments before every line
+      .replace(/\r?(^|\n)(?!$)/g, '$1-- ')
+      // Eliminate trailing spaces
+      .replace(/ +\n/g, '\n');
     return [getRunSqlQuery(comment, source)];
   }
   // all other errors